# swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: TelPro - FileMaker App (Hive)
# during dev, should point to your local machine
#host: 52.34.207.5:5058
host: localhost:5000
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /api/addPurchaseOrder:
    # binds a127 app logic to a route
    x-swagger-router-controller: purchaseorder_ctrl
    post:
      tags: ['Purchase Order Module']
      description: Add purchase order details
      summary: Add purchase order details
      # used as the method name of the controller
      operationId: addPurchaseOrder
      parameters:
        - name: body
          in: body
          description: Add PO
          schema:
           $ref: '#/definitions/addPurchaseOrder'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/updatePurchaseOrder:
    # binds a127 app logic to a route
    x-swagger-router-controller: purchaseorder_ctrl
    post:
      tags: ['Purchase Order Module']
      description: update purchase order details
      summary: update purchase order details
      # used as the method name of the controller
      operationId: updatePurchaseOrder
      parameters:
        - name: body
          in: body
          description: update purchase order
          schema:
           $ref: '#/definitions/updatePurchaseOrder'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/getPODetail:
    # binds a127 app logic to a route
    x-swagger-router-controller: purchaseorder_ctrl
    post:
      tags: ['Purchase Order Module']
      description: get  purchase order detail 
      summary: get purchase order detail
      # used as the method name of the controller
      operationId: getPODetail
      parameters:
        - name: body
          in: body
          description: get purchase order detail
          schema:
           $ref: '#/definitions/deletePO'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/getPOlist:
    # binds a127 app logic to a route
    x-swagger-router-controller: purchaseorder_ctrl
    post:
      tags: ['Purchase Order Module']
      description: get  purchase order list 
      summary: get purchase order list
      # used as the method name of the controller
      operationId: getPOlist
      parameters:
        - name: body
          in: body
          description: get purchase order list
          schema:
           $ref: '#/definitions/getPOlist'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/deletePO:
    # binds a127 app logic to a route
    x-swagger-router-controller: purchaseorder_ctrl
    post:
      tags: ['Purchase Order Module']
      description: delete purchase order 
      summary: delete purchase order 
      # used as the method name of the controller
      operationId: deletePO
      parameters:
        - name: body
          in: body
          description: delete purchase order
          schema:
           $ref: '#/definitions/deletePO'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/deletePOAttach:
    # binds a127 app logic to a route
    x-swagger-router-controller: purchaseorder_ctrl
    post:
      tags: ['Purchase Order Module']
      description: delete purchase order  attachment
      summary: delete purchase order  attachment
      # used as the method name of the controller
      operationId: deletePOAttach
      parameters:
        - name: body
          in: body
          description: delete purchase order attachment
          schema:
           $ref: '#/definitions/deletePOAttach'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/addAttachmentFile:
    # binds a127 app logic to a route
    x-swagger-router-controller: purchaseorder_ctrl
    post:
      tags: ['Purchase Order Module']
      description: add attachment for purchaseorder
      summary: add attachment for purchaseorder
      # used as the method name of the controller
      operationId: addAttachmentFile
      consumes:
        - multipart/form-data
        - application/x-www-form-urlencoded
        - binary
      parameters:
        - in: formData
          name: id
          description: PurchaseOrder Id
          type: string
          required: true                
        - in: formData
          name: description
          description: description for attachment 
          type: string
          required: true                
        - in: formData
          name: file
          description: add attachment for purchaseorder
          type: file
          required: true                      
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /api/getpurchaseOrderNo:
    # binds a127 app logic to a route
    x-swagger-router-controller: purchaseorder_ctrl
    post:
      tags: ['Purchase Order Module']
      description: get estimate number 
      summary:  get estimate number 
      # used as the method name of the controller
      operationId: getpurchaseOrderNo
      parameters:
        - name: body
          in: body
          description: get purchaseOrderNo 
          schema:
           $ref: '#/definitions/getpurchaseOrderNo'
          required: true 
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  appResponse:
    required:
      - code
    properties:
      code:
        type: number
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  addPurchaseOrder:
    type: object
    properties:  
      companyId:
        type: string
      contactId:
        type: string
      customerId:
        type: string
      projectId:
        type: string
      vendor:
        type: string
      vendorAddress1:
        type: string
      vendorAddress2:
        type: string
      vendorState:
        type: string
      vendorCity:
        type: string
      vendorZip:
        type: string
      shippingAddress1:
        type: string
      shippingAddress2:
        type: string
      shippingState:
        type: string
      shippingCity:
        type: string
      shippingZip:
        type: string
      requestedBy:
        type: string
      approvedBy:
        type: string
      shipVia:
        type: string
      statusId:
        type: number
      approvalStatusId:
        type: number
      createdDate:
        type: string
      approvedDate:
        type: string
      uponReceipt:
        type: string
      title:
        type: string
      trackingNumber:
        type: string
      shipDate:
        type: string
      notes:
        type: string
  deletePO:
    type: object
    properties:  
      poId:
        type: string
  deletePOAttach:
    type: object
    properties:  
      attachmentId:
        type: string
  getPOlist:
    type: object
    properties:  
      companyId:
        type: string
  updatePurchaseOrder:
    type: object
    properties:  
      companyId:
        type: string
      poId:
        type: string
      vendor:
        type: string
      vendorAddress1:
        type: string
      vendorAddress2:
        type: string
      vendorState:
        type: string
      vendorCity:
        type: string
      vendorZip:
        type: string
      shippingAddress1:
        type: string
      shippingAddress2:
        type: string
      shippingState:
        type: string
      shippingCity:
        type: string
      shippingZip:
        type: string
      requestedBy:
        type: string
      approvedBy:
        type: string
      shipVia:
        type: string
      trackingNumber:
        type: string
      statusId:
        type: number
      approvalStatusId:
        type: number
      createdDate:
        type: string
      approvedDate:
        type: string
      uponReceipt:
        type: string
      title:
        type: string
      shipDate:
        type: string
      notes:
        type: string
...
  getpurchaseOrderNo:
    type: object
    properties:
      companyId:
        type: string      
...